requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
# Pin motor to version compatible with Azure Cosmos DB MongoDB API (wire version 6)
# motor 3.3.x uses pymongo 4.6.x which requires MongoDB 4.2+ (wire version 8+)
# motor 3.1.x uses pymongo 4.3.x which supports wire version 6
motor>=3.1.0,<3.3.0
pymongo>=4.3.0,<4.6.0
python-dotenv>=1.0.0
pydantic[email]>=2.4.0
pydantic-settings>=2.0.0

# Authentication
PyJWT>=2.8.0

# HTTP client
aiohttp>=3.9.0

# Dapr SDK for pub/sub and service invocation
dapr>=1.16.0

# System monitoring
psutil>=5.9.0

# Azure Monitor / Application Insights (OpenTelemetry)
azure-monitor-opentelemetry==1.6.4
opentelemetry-instrumentation-fastapi==0.49b2
